# ggml-cuda : add rope f16, restore performance with parallel decoding