With this change, all the offloaded leafs are assumed to be inputs and copied to VRAM. This allows offloading `KQ_mask` and simplifies the logic for offloading `KQ_pos`. Replacing a conditional in the RoPE kernel with a template parameter as suggested by @JohannesGaessler also improves performance.

Performance compared to master:

Device 0: NVIDIA GeForce RTX 3090 Ti, compute capability 8.6
| model                          |       size | backend    | test       |       master t/s |           PR t/s | speedup |
| ------------------------------ | ---------: | ---------- | ---------- | ---------------: | ---------------: | ------: |
| llama 7B mostly Q2_K          |   2.63 GiB | CUDA       | pp 512     |  1805.63 ± 10.64 |   1828.37 ± 5.60 | 1.01259 |
| llama 7B mostly Q3_K - Small  |   2.75 GiB | CUDA       | pp 512     |   1907.34 ± 5.34 |   1932.17 ± 5.03 | 1.01301 |
| llama 7B mostly Q3_K - Medium |   3.07 GiB | CUDA       | pp 512     |   1997.51 ± 2.89 |   2017.52 ± 3.51 | 1.01001 |
| llama 7B mostly Q3_K - Large  |   3.35 GiB | CUDA       | pp 512     |   1928.15 ± 4.06 |   1951.13 ± 4.13 | 1.01191 |
| llama 7B mostly Q4_0          |   3.56 GiB | CUDA       | pp 512     |   2404.71 ± 3.79 |   2439.23 ± 5.24 | 1.01435 |
| llama 7B mostly Q4_K - Small  |   3.59 GiB | CUDA       | pp 512     |   2202.34 ± 3.30 |   2226.35 ± 5.70 | 1.01090 |
| llama 7B mostly Q4_K - Medium |   3.80 GiB | CUDA       | pp 512     |   2207.27 ± 3.36 |  2229.99 ± 11.03 | 1.01029 |
| llama 7B mostly Q4_1          |   3.95 GiB | CUDA       | pp 512     |   2102.01 ± 3.47 |   2123.36 ± 3.03 | 1.01015 |
| llama 7B mostly Q5_0          |   4.33 GiB | CUDA       | pp 512     |   2172.19 ± 4.07 |   2202.96 ± 5.92 | 1.01416 |
| llama 7B mostly Q5_K - Small  |   4.33 GiB | CUDA       | pp 512     |   2035.79 ± 4.42 |   2059.83 ± 5.22 | 1.01180 |
| llama 7B mostly Q5_K - Medium |   4.45 GiB | CUDA       | pp 512     |   2057.20 ± 4.03 |  2065.73 ± 24.31 | 1.00414 |
| llama 7B mostly Q5_1          |   4.72 GiB | CUDA       | pp 512     |   1931.67 ± 2.74 |   1948.27 ± 4.01 | 1.00859 |
| llama 7B mostly Q6_K          |   5.15 GiB | CUDA       | pp 512     |   2107.53 ± 1.66 |   2133.80 ± 7.40 | 1.01246 |
| llama 7B mostly Q8_0          |   6.67 GiB | CUDA       | pp 512     |   2353.88 ± 3.43 |   2386.22 ± 9.66 | 1.01373 |
| llama 7B mostly F16           |  12.55 GiB | CUDA       | pp 512     |   1659.42 ± 3.10 |   1679.42 ± 2.03 | 1.01205 |
| llama 7B mostly Q2_K          |   2.63 GiB | CUDA       | tg 128     |    105.91 ± 0.33 |    105.08 ± 1.66 | 0.99216 |
| llama 7B mostly Q3_K - Small  |   2.75 GiB | CUDA       | tg 128     |    101.85 ± 0.14 |    100.75 ± 1.97 | 0.98919 |
| llama 7B mostly Q3_K - Medium |   3.07 GiB | CUDA       | tg 128     |    108.07 ± 0.30 |    107.95 ± 0.90 | 0.99888 |
| llama 7B mostly Q3_K - Large  |   3.35 GiB | CUDA       | tg 128     |    105.34 ± 0.25 |    105.65 ± 0.36 | 1.00294 |
| llama 7B mostly Q4_0          |   3.56 GiB | CUDA       | tg 128     |    131.00 ± 0.30 |    130.90 ± 0.85 | 0.99923 |
| llama 7B mostly Q4_K - Small  |   3.59 GiB | CUDA       | tg 128     |    119.67 ± 0.23 |    120.00 ± 0.42 | 1.00275 |
| llama 7B mostly Q4_K - Medium |   3.80 GiB | CUDA       | tg 128     |    114.40 ± 0.14 |    114.15 ± 0.54 | 0.99781 |
| llama 7B mostly Q4_1          |   3.95 GiB | CUDA       | tg 128     |    124.94 ± 0.04 |    124.71 ± 0.67 | 0.99815 |
| llama 7B mostly Q5_0          |   4.33 GiB | CUDA       | tg 128     |    112.05 ± 0.43 |    111.67 ± 0.34 | 0.99660 |
| llama 7B mostly Q5_K - Small  |   4.33 GiB | CUDA       | tg 128     |    110.85 ± 0.11 |    111.08 ± 0.69 | 1.00207 |
| llama 7B mostly Q5_K - Medium |   4.45 GiB | CUDA       | tg 128     |    107.38 ± 0.06 |    106.77 ± 0.57 | 0.99431 |
| llama 7B mostly Q5_1          |   4.72 GiB | CUDA       | tg 128     |    108.43 ± 0.23 |    108.22 ± 0.78 | 0.99806 |
| llama 7B mostly Q6_K          |   5.15 GiB | CUDA       | tg 128     |     92.17 ± 0.07 |     91.27 ± 0.35 | 0.99023 |
| llama 7B mostly Q8_0          |   6.67 GiB | CUDA       | tg 128     |     88.62 ± 0.04 |     87.79 ± 0.48 | 0.99063 |
| llama 7B mostly F16           |  12.55 GiB | CUDA       | tg 128     |     56.14 ± 0.07 |     55.90 ± 0.16 | 0.99572 |