@ggerganov 2023-09-20
I'm testing this branch with the following command to fill-up the context and the results become incoherent after the first rope shift:

```
make -j && ./bin/main -m ../models/llama-7b/ggml-model-q8_0.gguf -p "I believe the meaning of life is" --ignore-eos -c 256 -n -1 -t 8 -ngl 35
```

It works OK with `-ngl 34`.
The following patch helps to see when the K-cache shifts occur, for easier debugging:

```diff
diff --git a/examples/main/main.cpp b/examples/main/main.cpp
index 1ed543c..2bf7628 100644
--- a/examples/main/main.cpp
+++ b/examples/main/main.cpp
@@ -502,6 +502,8 @@ int main(int argc, char ** argv) {
                 const int n_left    = n_past - params.n_keep - 1;
                 const int n_discard = n_left/2;
 
+                printf("\n\n XXXXXXXXXXXXXXXXXXXXXX\n\n");
+
                 LOG("context full, swapping: n_past = %d, n_left = %d, n_ctx = %d, n_keep = %d, n_discard = %d\n",
                     n_past, n_left, n_ctx, params.n_keep, n_discard);
 ```

<details>

```
llm_load_tensors: ggml ctx size =    0,09 MB
llm_load_tensors: using CUDA for GPU acceleration
llm_load_tensors: mem required  =  132,91 MB (+  128,00 MB per state)
llm_load_tensors: offloading 32 repeating layers to GPU
llm_load_tensors: offloading non-repeating layers to GPU
llm_load_tensors: offloading v cache to GPU
llm_load_tensors: offloading k cache to GPU
llm_load_tensors: offloaded 35/35 layers to GPU
llm_load_tensors: VRAM used: 6824 MB
...................................................................................................
llama_new_context_with_model: kv self size  =  128,00 MB
llama_new_context_with_model: compute buffer total size =   36,72 MB
llama_new_context_with_model: VRAM scratch buffer: 35,25 MB

system_info: n_threads = 8 / 64 | AVX = 1 | AVX2 = 1 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 1 | NEON = 0 | ARM_FMA = 0 | F16C = 1 | FP16_VA = 0 | WASM_SIMD = 0 | BLAS = 1 | SSE3 = 1 | SSSE3 = 1 | VSX = 0 | 
sampling: repeat_last_n = 64, repeat_penalty = 1,100000, presence_penalty = 0,000000, frequency_penalty = 0,000000, top_k = 40, tfs_z = 1,000000, top_p = 0,950000, typical_p = 1,000000, temp = 0,800000, mirostat = 0, mirostat_lr = 0,100000, mirostat_ent = 5,000000
generate: n_ctx = 256, n_batch = 512, n_predict = -1, n_keep = 0


 I believe the meaning of life is to love and serve others. The ultimate goal is becoming a servant-leader, and the best way to do that is through service and action.
As a young child in Vietnam, I had many friends who were not as fortunate as I was because of our war situation. In my teenage years, this realization made me feel guilty for being one of the lucky ones. My family fled to the United States when I was 14 years old and I’ve been living in the USA ever since. As a refugee myself, I understand the struggles of other refugees better than others because of my personal experience.
Teaching English is my passion, and it’s what keeps me inspired to serve. I started volunteering for a local nonprofit organization teaching English as a second language in 2015, and it has been the most rewarding and fulfilling part of my life since then. It helps me feel happy, proud, and productive!
My experience with refugees is not limited to students from Vietnam only; I have taught students from different countries including: Afghanistan, Burma (Myanmar), Bhutan, Laos,

 XXXXXXXXXXXXXXXXXXXXXX

 I love teaching English as a volunteer for nearly three years in the Portland Learning Center. As an ESL teacher at the same center since 2015; it has enriched my life and career-wise, even more than ever! Every time when I see them succeeding and improving their lives is very important to me teaching English as a volunteering for over five years ago – some places such as Sudanese, but also Syria, Iraq, Somalia, Congo, Afghanistan, Iran, Pakistan, Nepal, Vietnam, Eritia (Laos, Mynamia, Russia

 XXXXXXXXXXXXXXXXXXXXXX

, have learn about culture and have helped students. I’s now. The center in Seattle with a volunteer. My job is very same time as 16. When i, given me with so much I am able to my students are improving their lives. Thankfully!
and progressing, and also; because they improve my lives. Now.
I, I have as well. Thailand, Somalia, Kenya, Burma, Sudan, Sri Laba, and Sudh, Syria, Bhutan, Nepali, and Iraq and China, India, Tibet, Y

 XXXXXXXXXXXXXXXXXXXXXX

v I, their 1 have also been working as a year, and have a year-I’ve. I as.
Bes. I have also am I the, am happy for me in. Thaning.
Kathing and with a my English is much a. Also, and, I am volunteered at, soy, a have worked at afr.
Sudan, and Rwia, Eritisnaa and I have worked for. The work for 10, Bhutan.
Viety. e. and Vietnand I have and

 XXXXXXXXXXXXXXXXXXXXXX

 5, I the the last 6 to as an the same that they have work in, a to be a and as a
BH and the 3 and my, an as the
My havee work have tto be.
I am I, I
The are work. asI and work,
I have a I am with the ean and India, am
I am and in India. have I’ve worked an me and
have I
I am my have a, I have 1 I, am. am, I have.
I be the
am have I,

 XXXXXXXXXXXXXXXXXXXXXX

 in
inH, and work, asI work
The as. I have
be the last year,
I I, I h1, last
5,
I. I the work. The same. an, be the. I am a have the
My 6,be the,
to 7.
```
</details>