@slaren 2023-09-20
Should be fixed now. Unfortunately, the fix also makes the copies of K_pos, K_mask and K_shift synchronous, because they may be overwritten by other operations once the computation start (probably the `get_rows`, which still runs on the CPU). It shouldn't affect performance too much, but it should be fixed in the future.